feat : changed float to in in watcher

+++ src/Viserio/Component/Filesystem/Filesystem.php
@@ -503,7 +503,7 @@ class Filesystem extends SymfonyFilesystem implements FilesystemContract, Watche
     /**
      * {@inheritdoc}
      */
-    public function watch($path, callable $callback, ?float $timeout = null): void
+    public function watch($path, callable $callback, ?int $timeout = null): void
     {
         if (\extension_loaded('inotify')) {
             $watcher = new INotifyWatcher();
